# openapi.yaml
openapi: 3.1.1
info:
  title: API First
  version: 1.0.0
  description: API First

servers:
  - url: http://localhost:3000

components:
  schemas:
    HelloResponse:
      type: object
      properties:
        message:
          type: string

    UserInput:
      type: object
      required:
        - name
        - age
        - email
      properties:
        name:
          type: string
          minLength: 2
        age:
          type: integer
          minimum: 18
        email:
          type: string
          format: email
    User:
      allOf:
        - $ref: "#/components/schemas/UserInput"
        - type: object
          properties:
            id:
              type: string
paths:
  /hello:
    get:
      deprecated: false
      summary: Hello World
      responses:
        "200":
          description: A successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelloResponse"
  /users:
    post:
      summary: Crea un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      summary: "Obtener un usuario por su ID"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      deprecated: false
      summary: Actualizar un usuario por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string